Integration and unit tests; some refactoring

+++ backend/src/controllers/messagesController.js
@@ -1,13 +1,11 @@
+/*
+    TODO: Revisit to trim the fat here :(
+ */
+
 const mongoose = require('mongoose')
 const Message = require('../db/models/Message')
-const Utils = require('../utils/checkPalindrome')
-
-const idValidation = (req) => {
-    const { id: _id } = req
-    if (!mongoose.Types.ObjectId.isValid(_id))
-        return res.status(404).send('Not a valid ID')
-    return _id
-}
+const { isPalindrome } = require('../utils/palindrome')
+const { isValidId } = require('../utils/validator')
 
 const constructMessage = async (req) => {
     const message = new Message({
@@ -16,8 +14,8 @@ const constructMessage = async (req) => {
     })
     if (req.body.message !== '') {
         message.message = req.body.message
-        const isPalindrome = await Utils.checkPalindrome(req.body.message)
-        message.palindrome = isPalindrome
+        const palindrome = await isPalindrome(req.body.message)
+        message.palindrome = palindrome
     }
     return message
 }
@@ -32,8 +30,9 @@ const getMessages = async (req, res) => {
 }
 
 const getMessage = async (req, res) => {
-    const _id = await idValidation(req.params)
     try {
+        const { id: _id } = req.params
+        if (await !isValidId(_id)) return res.status(404).send('Not a valid ID')
         res.json(await Message.findById(_id))
     } catch (err) {
         res.json({ error: err.message })
@@ -51,8 +50,9 @@ const addMessage = async (req, res) => {
 }
 
 const updateMessage = async (req, res) => {
-    const _id = await idValidation(req.params)
     try {
+        const { id: _id } = req.params
+        if (await !isValidId(_id)) return res.status(404).send('Not a valid ID')
         const message = await constructMessage(req)
         const updatedMessage = await Message.updateOne(
             { _id: _id },
@@ -70,8 +70,9 @@ const updateMessage = async (req, res) => {
 }
 
 const removeMessage = async (req, res) => {
-    const _id = await idValidation(req.params)
     try {
+        const { id: _id } = req.params
+        if (await !isValidId(_id)) return res.status(404).send('Not a valid ID')
         const removedMessage = await Message.findByIdAndDelete({
             _id: _id,
         })
